requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
crispy-bootstrap5==2025.4
Django==5.1.7
django-bootstrap5==25.1
django-crispy-forms==2.4
Flask==3.1.0
frozenlist==1.6.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
multidict==6.4.3
pillow==11.1.0
propcache==0.3.1
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
twilio==9.6.0
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.20.0
